DeadlinesController: - implemented authorization on some methods

+++ app/Http/Controllers/API/DeadlinesController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\API;
 
 use App\Deadline;
+use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 use phpDocumentor\Reflection\Types\Integer;
 
@@ -30,12 +32,18 @@ class DeadlinesController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = Validator::make($request->all(), Deadline::$validation_rules);
-        if($validator->fails()){
-            return response()->json($validator->errors(), 400);
+        $current_user = User::find(Auth::user()->id);
+        if($current_user->hasRole('Admin')) {
+            $validator = Validator::make($request->all(), Deadline::$validation_rules);
+            if ($validator->fails()) {
+                return response()->json($validator->errors(), 400);
+            }
+            $deadline = Deadline::create($request->all());
+            return response()->json($deadline, 201);
         }
-        $deadline = Deadline::create($request->all());
-        return response()->json($deadline, 201);
+        else{
+            return response()->json(null, 403);
+        }
     }
 
     /**
@@ -59,13 +67,19 @@ class DeadlinesController extends Controller
      */
     public function update(Request $request, Deadline $deadline)
     {
-        $validator = Validator::make($request->all(), Deadline::$validation_rules);
-        if($validator->fails()){
-            return response()->json($validator->errors(), 400);
-        }
+        $current_user = User::find(Auth::user()->id);
+        if($current_user->hasRole('Admin')) {
+            $validator = Validator::make($request->all(), Deadline::$validation_rules);
+            if ($validator->fails()) {
+                return response()->json($validator->errors(), 400);
+            }
 
-        $deadline->update($request->all());
-        return response()->json($deadline, 200);
+            $deadline->update($request->all());
+            return response()->json($deadline, 200);
+        }
+        else{
+            return response()->json(null, 403);
+        }
     }
 
     /**
@@ -76,11 +90,17 @@ class DeadlinesController extends Controller
      */
     public function destroy($id)
     {
-        $deadline = Deadline::find($id);
-        if(is_null($deadline)){
-            return response()->json(null, 404);
+        $current_user = User::find(Auth::user()->id);
+        if($current_user->hasRole('Admin')) {
+            $deadline = Deadline::find($id);
+            if (is_null($deadline)) {
+                return response()->json(null, 404);
+            }
+            $deadline->delete();
+            return response()->json(null, 204);
         }
-        $deadline->delete();
-        return response()->json(null, 204);
+        else{
+            return response()->json(null, 403);
+        }
     }
 }
